@@ -141,8 +142,8 @@
     "tap-spot": "^1.1.1",
     "textlint": "^11.8.2",
     "textlint-rule-heading": "^1.0.10",
-    "unist-util-visit": "^2.0.3",
-    "webpack": "^5.52.1",
+    "unist-util-visit": "^4.0.0",
+    "webpack": "^5.53.0",
     "webpack-bundle-analyzer": "^4.4.2",
     "webpack-cli": "^4.8.0",
     "webpack-dev-server": "^4.2.1",
