@@ -1,12 +1,16 @@
-const _ = require('lodash');
-const fs = require('fs');
-const path = require('path');
-const { promisify } = require('util');
-const mkdirp = require('mkdirp');
-const fetch = require('node-fetch');
+import _ from 'lodash';
+import fs from 'fs';
+import path from 'path';
+import { promisify } from 'util';
+import mkdirp from 'mkdirp';
+import fetch from 'node-fetch';
+import { fileURLToPath } from 'url';
 
-const yamlHeadmatter = require('./yaml-headmatter.js');
-const processReadme = require('./process-readme.js');
+import yamlHeadmatter from './yaml-headmatter.mjs';
+import processReadme from './process-readme.mjs';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
 
 const writeFile = promisify(fs.writeFile);
 const rename = promisify(fs.rename);
