@@ -1,17 +1,31 @@
-const path = require('path');
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-const webpack = require('webpack');
-const h = require('hastscript');
-const remarkResponsiveTable = require('./src/remark-plugins/remark-responsive-table/remark-responsive-table.js');
+import path from 'path';
+import MiniCssExtractPlugin from 'mini-css-extract-plugin';
+import webpack from 'webpack';
+import h from 'hastscript';
+import remarkResponsiveTable from './src/remark-plugins/remark-responsive-table/remark-responsive-table.mjs';
+import gfm from 'remark-gfm';
+import slug from './src/remark-plugins/docschina-remark-slugger/index.mjs';
+import cleanup from './src/remark-plugins/remark-cleanup-readme/index.mjs';
+import aside from './src/remark-plugins/remark-custom-asides/index.mjs';
+import autolink from 'remark-autolink-headings';
+import refractor from 'remark-refractor';
+import frontmatter from 'remark-frontmatter';
+import { createRequire } from 'module';
+import remarkEmoji from 'remark-emoji';
+import { fileURLToPath } from 'url';
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+
+const require = createRequire(import.meta.url);
+
 const mdPlugins = [
-  require('remark-gfm'),
-  require('./src/remark-plugins/docschina-remark-slugger/index.js'),
-  // require('remark-attr'),
+  gfm,
+  slug,
   remarkResponsiveTable,
-  require('remark-emoji'),
-  require('./src/remark-plugins/remark-cleanup-readme/index.js'),
+  remarkEmoji,
+  cleanup,
   [
-    require('./src/remark-plugins/remark-custom-asides/index.js'),
+    aside,
     {
       mapping: {
         'T>': 'tip',
@@ -21,7 +35,7 @@ const mdPlugins = [
     },
   ],
   [
-    require('remark-autolink-headings'),
+    autolink,
     {
       behavior: 'append',
       content() {
@@ -29,10 +43,10 @@ const mdPlugins = [
       },
     },
   ],
-  require('remark-refractor'),
+  refractor,
 ];
 
-module.exports = ({ ssg = false }) => ({
+export default ({ ssg = false }) => ({
   context: path.resolve(__dirname, './src'),
   cache: {
     type: 'filesystem',
@@ -56,7 +70,7 @@ module.exports = ({ ssg = false }) => ({
           {
             loader: '@mdx-js/loader',
             options: {
-              remarkPlugins: [...mdPlugins, [require('remark-frontmatter')]],
+              remarkPlugins: [...mdPlugins, [frontmatter]],
             },
           },
         ],
