@@ -1,18 +1,20 @@
-import remark from 'remark';
+import { remark } from 'remark';
+import remarkHtml from 'remark-html';
+import Plugin from './index.mjs';
 describe('customize blockquote', () => {
   it('should transform W> into aside of warning', () => {
     remark()
-      .use(require('./index.js'), {
+      .use(Plugin, {
         mapping: {
           'W>': 'warning',
         },
       })
-      .use(require('remark-html'))
+      .use(remarkHtml)
       .process(
         `
 W> hello world
 `,
-        function (err, { contents }) {
+        function (err, { value: contents }) {
           expect(err).toBeNull();
           expect(contents).toContain('<aside class="warning"');
           expect(contents).toContain('<h6 class="warning__prefix"');
