@@ -1,9 +1,14 @@
-import remark from 'remark';
+import { describe, expect } from '@jest/globals';
+
+import { remark } from 'remark';
+import remarkHtml from 'remark-html';
+import remarkGfm from 'remark-gfm';
+import remarkResponsiveTable from './remark-responsive-table.mjs';
 describe('responsive table', () => {
   const processor = remark()
-    .use(require('remark-gfm'))
-    .use(require('./remark-responsive-table.js'))
-    .use(require('remark-html'));
+    .use(remarkGfm)
+    .use(remarkResponsiveTable)
+    .use(remarkHtml);
   it('should add data-th', () => {
     processor.process(
       `
@@ -12,7 +17,7 @@ describe('responsive table', () => {
 | baz |  |
 | sam | chen |
       `,
-      (error, { contents }) => {
+      (error, { value: contents }) => {
         expect(error).toBeNull();
         expect(contents).toContain('data-th="foo"');
         expect(contents).toContain('data-th="bar"');
@@ -29,7 +34,7 @@ describe('responsive table', () => {
 | baz |  |
 | sam | chen |
       `,
-      (error, { contents }) => {
+      (error, { value: contents }) => {
         expect(error).toBeNull();
         expect(contents).toMatchSnapshot();
       }
