@@ -1,4 +1,4 @@
-const excludedLoaders = [
+export const excludedLoaders = [
   'webpack-contrib/config-loader',
   'webpack-contrib/eslint-loader',
   'webpack-contrib/transform-loader',
@@ -25,7 +25,7 @@ const excludedLoaders = [
   'webpack-contrib/istanbul-instrumenter-loader',
   'webpack-contrib/worker-loader',
 ];
-const excludedPlugins = [
+export const excludedPlugins = [
   'webpack-contrib/component-webpack-plugin',
   'webpack-contrib/extract-text-webpack-plugin',
   'webpack-contrib/i18n-webpack-plugin',
@@ -34,8 +34,3 @@ const excludedPlugins = [
   'webpack-contrib/zopfli-webpack-plugin',
   'webpack-contrib/closure-webpack-plugin',
 ];
-
-module.exports = {
-  excludedLoaders,
-  excludedPlugins,
-};
